Add pirogue.gpg file.
This is a copy of pirogue-archive-keyring.gpg as shipped in version 0.1
of the new pirogue-archive-keyring package. We're publishing the file,
alongside instructions on how to download it, under a different name to
make it easy on users: if and when both differ, users might get a prompt
asking them what they want to do with that file (defaulting to keeping
the local file, as opposed to deploying the new one coming from the
package), which is likely to generate confusion, and might end up with
users not getting an updated keyring.

This means the published file is only useful when bootstrapping the
environment, the pirogue-archive-keyring package can take over. Once the
transition period is over, we might even include some postinst code to
to keep /etc/apt/trusted.gpg.d/ clean:
 - pirogue.asc goes away (historical file);
 - pirogue.gpg goes away (this new file);
 - irogue-archive-keyring.gpg stays (package-maintained).

Meanwhile, in this repository, let's keep the historical `Key.gpg`
(which is actually an ASCII-armored export of a single key, deployed via
curl as `pirogue.asc`), until everything is updated to use the new file.