Replace `curl -o` with `wget -O`.
We're already making sure wget is installed in Pi images, and it's
also the one that can be expected to be installed in virtually all
Debian environments, as it's `Priority: standard` while curl is
only `Priority: optional`.

What matters is what gets documented on the website ultimately, but
it makes sense to improve consistency on the PPA side already.